Module 4 - Inserting Updating and Deleting Data/requirements.txt
django
djangorestframework
psycopg2-binary
drf-spectacular